ZoneRate as interface #1053
