+++ commercetools-models/src/main/java/io/sphere/sdk/shippingmethods/ZoneRate.java
@@ -1,8 +1,7 @@
 package io.sphere.sdk.shippingmethods;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.models.Reference;
 import io.sphere.sdk.models.Referenceable;
 import io.sphere.sdk.zones.Zone;
@@ -15,33 +14,14 @@ import java.util.List;
  * @see ShippingMethod#getZoneRates()
  * @see ShippingMethodDraft#getZoneRates()
  */
-public final class ZoneRate extends Base {
-    private final Reference<Zone> zone;
-    private final List<ShippingRate> shippingRates;
-
-    @JsonCreator
-    private ZoneRate(final Reference<Zone> zone, final List<ShippingRate> shippingRates) {
-        this.zone = zone;
-        this.shippingRates = shippingRates;
-    }
-
-    /**
-     * Creates a new {@link ZoneRate}.
-     *
-     * @param zone zone for the rates to apply
-     * @param shippingRates costs for different currencies
-     * @return zone rate
-     */
+@JsonDeserialize(as = ZoneRateImpl.class)
+public interface ZoneRate {
     @JsonIgnore
-    public static ZoneRate of(final Referenceable<Zone> zone, final List<ShippingRate> shippingRates) {
-        return new ZoneRate(zone.toReference(), shippingRates);
+    static ZoneRate of(final Referenceable<Zone> zone, final List<ShippingRate> shippingRates) {
+        return new ZoneRateImpl(zone.toReference(), shippingRates);
     }
 
-    public Reference<Zone> getZone() {
-        return zone;
-    }
+    Reference<Zone> getZone();
 
-    public List<ShippingRate> getShippingRates() {
-        return shippingRates;
-    }
+    List<ShippingRate> getShippingRates();
 }
